+++ packages/visualizations/src/components/Chart/utils.ts
@@ -1,4 +1,4 @@
-import type { Color } from '../types';
+import type { Color } from 'types';
 
 export const DEFAULT_GREY_COLOR = '#F0F0F0';
