+++ packages/visualizations/src/components/KpiCard/types.ts
@@ -1,5 +1,4 @@
-import type { Color, Source } from '../types';
-import type { Async } from '../../types';
+import type { Async, Color, Source } from 'types';
 
 export interface KpiCardOptions {
     title?: string;
