+++ packages/visualizations/src/components/Legend/CategoryLegend/Item/CategoryLegendItemSymbol.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
+    import type { CategoryItem } from 'types';
+    import { CATEGORY_ITEM_VARIANT } from '../../types';
     import BoxSymbol from '../Symbols/BoxSymbol.svelte';
     import LineSymbol from '../Symbols/LineSymbol.svelte';
     import CircleSymbol from '../Symbols/CircleSymbol.svelte';
     import ImageSymbol from '../Symbols/ImageSymbol.svelte';
-    import type { CategoryItem } from '../../types';
-    import { CATEGORY_ITEM_VARIANT } from '../../types';
 
     export let item: CategoryItem;
 </script>
