+++ packages/visualizations/src/components/Legend/CategoryLegend/Symbols/BoxSymbol.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-    import type { BoxCategoryItem } from '../../types';
+    import type { BoxCategoryItem } from 'types';
 
     export let item: BoxCategoryItem;
 </script>
