+++ packages/visualizations/src/components/Legend/ColorsLegend.svelte
@@ -3,9 +3,9 @@
 <script lang="ts">
     import { onDestroy } from 'svelte';
     import { debounce } from 'lodash';
-    import type { DataBounds, ColorScale } from '../types';
+    import type { DataBounds, ColorScale } from 'types';
+    import { defaultCompactLegendNumberFormat } from 'components/utils/formatter';
     import type { LegendVariant, LegendPositions } from './types';
-    import { defaultCompactLegendNumberFormat } from '../utils/formatter';
 
     // options to customize the component
     export let dataBounds: DataBounds;
