+++ packages/visualizations/src/components/Map/types.ts
@@ -1,8 +1,8 @@
 import type { Feature, FeatureCollection, Position, BBox } from 'geojson';
 import type { FillLayerSpecification, Popup, GestureOptions, LngLatBoundsLike } from 'maplibre-gl';
 import type { DebouncedFunc } from 'lodash';
-import type { LegendPositions } from '../Legend/types';
-import type { ColorScale, Color, Source } from '../types';
+import type { ColorScale, Color, Source } from 'types';
+import type { LegendPositions } from 'components/Legend/types';
 
 export interface ChoroplethOptions {
     /** Configuration for the color scale used to color the choropleth shapes. */
