+++ packages/visualizations/src/components/Map/utils.ts
@@ -4,15 +4,9 @@ import maplibregl, { ExpressionInputType, ExpressionSpecification } from 'maplib
 import { viewport } from '@placemarkio/geo-viewport';
 import type { Feature, FeatureCollection, Position, BBox } from 'geojson';
 import type { Scale } from 'chroma-js';
+import { assertUnreachable } from 'components/utils';
+import { Color, ColorScale, DataBounds, isGroupByForMatchExpression, ColorScaleTypes } from 'types';
 import { DEFAULT_COLORS } from './constants';
-import { assertUnreachable } from '../utils';
-import {
-    Color,
-    ColorScale,
-    DataBounds,
-    isGroupByForMatchExpression,
-    ColorScaleTypes,
-} from '../types';
 
 import type {
     ChoroplethDataValue,
