+++ packages/visualizations/src/components/Table/types.ts
@@ -1,7 +1,6 @@
-import type { DataFrame, Source } from '../types';
+import type { DataFrame, Source, Async } from 'types';
 import type { DATA_FORMAT } from './constants';
 import type { Pagination } from '../Pagination/types';
-import type { Async } from '../../types';
 
 type DataFormatKey = keyof typeof DATA_FORMAT;
 export type DataFormat = typeof DATA_FORMAT[DataFormatKey];
